# docs/specs/api-openapi.yaml
...
openapi: 3.1.0
info:
  title: PromptPotter Control-remote API
  version: 0.0.0-profile-minus-1
  summary: Closed inbound command set for the M12 Control-remote I/O kind.
  description: |
    The HTTP command surface for PromptPotter's interactive control plane.
    Every operation is a Control-remote command: an HTTP-ingressed mutation
    authored by a signed-in operator or signed-in client, appended to its
    target ledger (the per-cycle `.runtime/ledger.jsonl` or the workspace
    `.workspace/events.jsonl`, per command scope) as a `CommandRecord` by the
    sole `CommandDispatcher`.

    See `docs/adr/0001-m12-control-plane.md` for the permanent contract, the
    Profile gradient (A-E), and the 20-item security checklist that every
    handler is measured against.

    Conformance: handlers MUST reject any request that omits the
    `Idempotency-Key` header. State-mutating handlers MUST additionally
    reject any request that omits the `Expected-Version` header. Mismatch
    on `Expected-Version` returns 409 `version_conflict`.

  contact:
    name: PromptPotter maintainers
  license:
    name: See repository LICENSE
servers:
  - url: http://localhost:8001
    description: Local dev (uvicorn + auth-off, Stage 0 / Profile B default).
  - url: https://{install}.promptpotter.com
    description: Hosted hub (Profile D — placeholder, OIDC required).
    variables:
      install:
        default: example
        description: Per-install hostname slug.
...
paths:
...
  /datasets/{name}/cells:
    get:
      operationId: getDatasetCells
      summary: The measurement log of one scope — every cell (candidate × sample), served ranked.
      description: |
        One read behind every measurement view. `samples` in hard-sample rank,
        `candidates` chronological within a cycle, `cells` by candidate then
        walk order (the time series); a
        client groups by bucketing under those served orders and never
        re-sorts. Grouped by sample it IS the hard-sample leaderboard. The
        round still being measured is merged here, once. `candidate_id` /
        `round` / `status` narrow the rows — a preset, not a second read.
        Wire shape: `domain/cells.py::CellsResponse`.
      tags: [datasets]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: scope, in: query, schema: { type: string, enum: [cycle, campaign, dataset], default: dataset } }
        - { name: campaign_id, in: query, schema: { type: string } }
        - { name: cycle_id, in: query, schema: { type: string } }
        - { name: descend, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [info_gain, difficulty] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 50 } }
        - { name: max_unmeasured, in: query, schema: { type: integer, minimum: 0, maximum: 1000 } }
        - { name: candidate_id, in: query, schema: { type: string } }
        - { name: round, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [HIT, MISS, ERR, UNSC] } }
      responses:
        '200':
          description: The scope's ranked samples, its candidates and their cells.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: No such dataset, campaign or cycle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /datasets/{name}/cells/{run_id}/{sample_id}:
    get:
      operationId: getDatasetCell
      summary: One cell opened — its archive row assembled into a trace.
      description: |
        Addressed `(run_id, sample_id)`: the archive run the row was filed
        under and its sample. One span per pipeline node, in chain order —
        the input re-rendered from the run's own node config, tokens and
        seconds off the row, outputs attributed through the dataset's schema.
        Assembled at read time; nothing is stored for it. Wire shape:
        `domain/cells.py::Cell`.
      tags: [datasets]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: run_id, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: sample_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: The cell and its spans.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: No such dataset, no such run under it, or the run holds no cell for that sample.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error, message, error_id]
      additionalProperties: false
      properties:
        error:
          type: string
          description: |
            Stable error code. The whole API serializes errors to this one flat
            envelope (no `detail` wrapper) — every code below maps from a
            `PotterError` subclass at the single `main.py` handler. Control-plane
            specifics: `version_conflict` (409), `idempotency_key_missing` (400),
            `command_kind_unknown` (404), `slug_collision` (409),
            `machine_busy` (409), `cycle_busy` (409), `backend_unreachable` (503),
            `admission_gate_stuck` (503 — a process is wedged inside the
            machine-wide admission window, so occupancy cannot be read and
            nothing was started; retryable),
            `ingest_failed` / `origin_incomplete` / `pipeline_config_invalid`
            (422), `stored_config_invalid` (500). Generic families:
            `bad_request` (400), `unauthenticated`
            (401), `capability_denied` (403), `not_found` (404), `too_large`
            (413), `request_invalid` (422), `service_unavailable` (503),
            `internal_error` (500). New codes require a YAML update first.
          enum:
            - version_conflict
            - idempotency_key_missing
            - expected_version_missing
            - capability_denied
            - command_kind_unknown
            - command_target_not_found
            - payload_invalid
            - tenant_mismatch
            - slug_collision
            - machine_busy
            - cycle_busy
            - backend_unreachable
            - ingest_failed
            - origin_incomplete
            - pipeline_config_invalid
            - stored_config_invalid
            - nothing_to_replace
            - resolver_failed
            - backend_sync_failed
            - job_registry_unavailable
            - identity_not_initialised
            - provider_not_configured
            - provider_unknown
            - bad_request
            - unauthenticated
            - not_found
            - too_large
            - request_invalid
            - service_unavailable
            - internal_error
        message:
          type: string
          description: Human-readable explanation; never load-bearing for clients.
        error_id:
          type: string
          description: |
            Short hex handle stamped on EVERY error response and written to the
            same server log line (`api error [<id>] <method> <path> -> <status>
            <code>`; a 500 logs its traceback under the same handle). It is the
            one thing that ties an opaque client-side failure to its server-side
            cause, so a bug report quotes it instead of a wall-clock guess — the
            webapp surfaces it through Account → Copy diagnostics. Minted at the
            single `main.py::_error_response` seam; never reused, never
            client-supplied.
        details:
          type: object
          additionalProperties: true
          description: |
            Optional per-error structured detail.

            * `version_conflict` — `expected_version` (client sent) +
              `actual_version` (runner observed).
            * `slug_collision` (409 on `/datasets/ingest`) — `suggested_slug`
              carrying the smallest free `{slug}-{n}`, and `slug` (the colliding
              name) so the chat can offer "use existing / save as new".
            * `machine_busy` — **no launch command returns this.** A full box
              QUEUES rather than refuses, so `mint-campaign`, `start-checkin`,
              `start-run` and `fork-cycle` all accept and the launch waits its
              turn. The code survives for the terminal's `--no-wait`, which asks
              to be refused instead; it carries the holder presence record
              `holder_user`, `campaign_id`, `cycle_id`, `started_at`. What the
              webapp reads is `GET /machine-status` — occupancy, and the
              caller's own place in line.
            * `cycle_busy` (409 on `start-run`, `step-cycle` and
              `start-checkin`) — the targeted cycle already has an unfinished
              job (queued included), so a second launch would put two
              producers on one ledger. Carries that job: `job_id`, `status`,
              `holder_user`, `campaign_id`, `cycle_id`, `started_at`.
            * `backend_unreachable` (503 on any command that needs the
              backend reachable) — `backend_type` (connector name) and
              `backend_url` (configured URL) — the IDENTITY of what refused.
              The REMEDY is the connector probe's own and rides `message`: a
              surface renders that, never a sentence rebuilt from these two,
              which cannot say why.
            * `ingest_failed` (422 on `/datasets/ingest`) — `reason`
              ∈ `bad_csv` | `bad_json` | `empty` | `too_large` | `bad_slug` |
              `unsupported_format` | `hardened_blocked`. (Ingest is
              header-agnostic — it no longer rejects on column names; column
              identity is resolved on the draft, gated at mint. `hardened_blocked`
              is an XLSX upload refused while the server runs `HARDENED_MODE`.)
            * `origin_incomplete` (422 on `/commands/start-checkin`)
              — `gaps`: a list of `{field, reason, hint}` objects naming each
              origin field that still blocks mint. `reason` ∈ `unset` |
              `proposed_unconfirmed`. The deterministic origin-readiness
              checklist — not the operator — is the source of truth for
              completeness (see `docs/specs/roadmap.md`).
            * `stored_config_invalid` (500 on any read of a stored config the
              server itself persisted) — `path` (the workspace-relative file)
              and `reason` (what the parse or validation rejected, naming the
              offending key). Distinct from `pipeline_config_invalid` (422),
              which rejects config the *caller* supplied: this one means data on
              disk stopped satisfying the code that reads it — typically a knob
              dropped from `CampaignConfig` (`extra="forbid"`) while a file
              written earlier still names it. It is a 500 because the caller did
              nothing wrong and cannot fix it; the remedy is the
              `promptpotter restamp --apply` verb, which `deploy-linux/update.sh`
              runs on every deploy.
            (`internal_error` carries no `details`: its trace handle is the
            top-level `error_id`, which every error now has.)

    # ---- Read-surface shapes (the forest's cycle list, at any depth) ----
    # Mirrors promptpotter/presentation/api/routers/active.py::CycleListEntry
    # + CyclesResponse. GET /cycles serves it at every depth via ?descend=
    # (declared above) — top level, an L4 cycle's inner fan-out, or L5+.
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: pp_session
      description: |
        Opaque server-side session id (NOT a JWT). Issued by the OIDC
        middleware at Profile C after ID Token verification; per
        identity-foundation no-drift gate #2 the cookie itself carries no
        claims. At Profile B (auth-off), this scheme is declared but the
        Stage-0 default identity is returned unconditionally.